+++ internal/model.go
@@ -68,7 +68,7 @@ type NormalizationConstants struct {
 	MaxAmount            float32 `json:"max_amount"`
 	MaxInstallments      float32 `json:"max_installments"`
 	AmountVsAvgRatio     float32 `json:"amount_vs_avg_ratio"`
-	MaxMinutes           int     `json:"max_minutes"`
+	MaxMinutes           float32 `json:"max_minutes"`
 	MaxKm                float32 `json:"max_km"`
 	MaxTxCount24h        float32 `json:"max_tx_count_24h"`
 	MaxMerchantAvgAmount float32 `json:"max_merchant_avg_amount"`
